requirements.txt
fastapi==0.111.0
uvicorn==0.22.0
sqlalchemy==2.0.0
pydantic==1.10.0
langchain_community==0.2.19
langchain_chroma==0.2.2
pytest==8.3.5
qianfan==0.4.12.3
pymysql==1.1.1
firebase_admin==6.7.0
passlib==1.7.4
langgraph==0.059
alembic==1.15.2